refactor: rename files and move into movement directory
