+++ libs/cartesiandomain/movement/CMakeLists.txt
@@ -12,6 +12,7 @@ set(LIB_BINARY_DIR ${CMAKE_BINARY_DIR}/lib)
 
 # Add executables and create library target
 set(SOURCES
+  "add_supers_at_domain_top.cpp"
   "cartesian_transport_across_domain.cpp"
 )
 # must use STATIC (not(!) SHARED) lib for linking to executable if build is CUDA enabled with Kokkos
