+++ roughpaper/src/main_impl.hpp
@@ -31,12 +31,12 @@
 #include <string_view>
 
 #include "zarr/dataset.hpp"
+#include "cartesiandomain/movement/add_supers_at_domain_top.hpp"
 #include "cartesiandomain/movement/cartesian_movement.hpp"
-#include "cartesiandomain/add_supers_at_domain_top.hpp"
+#include "cartesiandomain/movement/cartesian_motion.hpp"
+#include "cartesiandomain/movement/null_boundary_conditions.hpp"
 #include "cartesiandomain/cartesianmaps.hpp"
-#include "cartesiandomain/cartesianmotion.hpp"
 #include "cartesiandomain/createcartesianmaps.hpp"
-#include "cartesiandomain/null_boundary_conditions.hpp"
 #include "coupldyn_fromfile/fromfile_cartesian_dynamics.hpp"
 #include "coupldyn_fromfile/fromfilecomms.hpp"
 #include "gridboxes/gridboxmaps.hpp"
